+++ test/assertions/runAssertions.test.ts
@@ -1,0 +1,434 @@
+/* eslint-disable @typescript-eslint/no-unused-vars, unused-imports/no-unused-imports */
+import dedent from 'dedent';
+import * as fs from 'fs';
+import * as path from 'path';
+import { runAssertion, runAssertions } from '../../src/assertions';
+import { validateXml, containsXml } from '../../src/assertions/xml';
+import { fetchWithRetries } from '../../src/fetch';
+import { OpenAiChatCompletionProvider } from '../../src/providers/openai/chat';
+import {
+  DefaultGradingJsonProvider,
+  DefaultEmbeddingProvider,
+} from '../../src/providers/openai/defaults';
+import { ReplicateModerationProvider } from '../../src/providers/replicate';
+import type {
+  Assertion,
+  ApiProvider,
+  AtomicTestCase,
+  ProviderResponse,
+  GradingResult,
+} from '../../src/types';
+import { TestGrader } from '../util/utils';
+
+jest.mock('../../src/redteam/remoteGeneration', () => ({
+  shouldGenerateRemote: jest.fn().mockReturnValue(false),
+}));
+
+jest.mock('proxy-agent', () => ({
+  ProxyAgent: jest.fn().mockImplementation(() => ({})),
+}));
+
+jest.mock('node:module', () => {
+  const mockRequire: NodeJS.Require = {
+    resolve: jest.fn() as unknown as NodeJS.RequireResolve,
+  } as unknown as NodeJS.Require;
+  return {
+    createRequire: jest.fn().mockReturnValue(mockRequire),
+  };
+});
+
+jest.mock('../../src/fetch', () => {
+  const actual = jest.requireActual('../../src/fetch');
+  return {
+    ...actual,
+    fetchWithRetries: jest.fn(actual.fetchWithRetries),
+  };
+});
+
+jest.mock('glob', () => ({
+  globSync: jest.fn(),
+}));
+
+jest.mock('fs', () => ({
+  readFileSync: jest.fn(),
+  promises: {
+    readFile: jest.fn(),
+  },
+}));
+
+jest.mock('../../src/esm');
+jest.mock('../../src/database', () => ({
+  getDb: jest.fn(),
+}));
+jest.mock('path', () => ({
+  ...jest.requireActual('path'),
+  resolve: jest.fn(jest.requireActual('path').resolve),
+  extname: jest.fn(jest.requireActual('path').extname),
+}));
+
+jest.mock('../../src/cliState', () => ({
+  basePath: '/base/path',
+}));
+jest.mock('../../src/matchers', () => {
+  const actual = jest.requireActual('../../src/matchers');
+  return {
+    ...actual,
+    matchesContextRelevance: jest
+      .fn()
+      .mockResolvedValue({ pass: true, score: 1, reason: 'Mocked reason' }),
+    matchesContextFaithfulness: jest
+      .fn()
+      .mockResolvedValue({ pass: true, score: 1, reason: 'Mocked reason' }),
+  };
+});
+
+const Grader = new TestGrader();
+
+describe('runAssertions', () => {
+  const test: AtomicTestCase = {
+    assert: [
+      {
+        type: 'equals',
+        value: 'Expected output',
+      },
+    ],
+  };
+
+  beforeEach(() => {
+    jest.resetModules();
+  });
+
+  afterEach(() => {
+    jest.clearAllMocks();
+  });
+
+  it('should pass when all assertions pass', async () => {
+    const output = 'Expected output';
+
+    const result: GradingResult = await runAssertions({
+      prompt: 'Some prompt',
+      provider: new OpenAiChatCompletionProvider('gpt-4o-mini'),
+      test,
+      providerResponse: { output },
+    });
+    expect(result).toMatchObject({
+      pass: true,
+      reason: 'All assertions passed',
+    });
+  });
+
+  it('should fail when any assertion fails', async () => {
+    const output = 'Actual output';
+
+    const result: GradingResult = await runAssertions({
+      prompt: 'Some prompt',
+      provider: new OpenAiChatCompletionProvider('gpt-4o-mini'),
+      test,
+      providerResponse: { output },
+    });
+    expect(result).toMatchObject({
+      pass: false,
+      reason: 'Expected output "Actual output" to equal "Expected output"',
+    });
+  });
+
+  it('should handle output as an object', async () => {
+    const output = { key: 'value' };
+
+    const result: GradingResult = await runAssertions({
+      prompt: 'Some prompt',
+      provider: new OpenAiChatCompletionProvider('gpt-4o-mini'),
+      test,
+      providerResponse: { output },
+    });
+    expect(result).toMatchObject({
+      pass: false,
+      reason: 'Expected output "{"key":"value"}" to equal "Expected output"',
+    });
+  });
+
+  it('should fail when combined score is less than threshold', async () => {
+    const result: GradingResult = await runAssertions({
+      prompt: 'Some prompt',
+      provider: new OpenAiChatCompletionProvider('gpt-4o-mini'),
+      test: {
+        threshold: 0.5,
+        assert: [
+          {
+            type: 'equals',
+            value: 'Hello world',
+            weight: 2,
+          },
+          {
+            type: 'contains',
+            value: 'world',
+            weight: 1,
+          },
+        ],
+      },
+      providerResponse: { output: 'Hi there world' },
+    });
+    expect(result).toMatchObject({
+      pass: false,
+      reason: 'Aggregate score 0.33 < 0.5 threshold',
+    });
+  });
+
+  it('should pass when combined score is greater than threshold', async () => {
+    const result: GradingResult = await runAssertions({
+      prompt: 'Some prompt',
+      provider: new OpenAiChatCompletionProvider('gpt-4o-mini'),
+      test: {
+        threshold: 0.25,
+        assert: [
+          {
+            type: 'equals',
+            value: 'Hello world',
+            weight: 2,
+          },
+          {
+            type: 'contains',
+            value: 'world',
+            weight: 1,
+          },
+        ],
+      },
+      providerResponse: { output: 'Hi there world' },
+    });
+    expect(result).toMatchObject({
+      pass: true,
+      reason: 'Aggregate score 0.33 ≥ 0.25 threshold',
+    });
+  });
+
+  describe('assert-set', () => {
+    const prompt = 'Some prompt';
+    const provider = new OpenAiChatCompletionProvider('gpt-4o-mini');
+
+    it('assert-set success', async () => {
+      const output = 'Expected output';
+      const test: AtomicTestCase = {
+        assert: [
+          {
+            type: 'assert-set',
+            assert: [
+              {
+                type: 'equals',
+                value: output,
+              },
+            ],
+          },
+        ],
+      };
+
+      const result: GradingResult = await runAssertions({
+        prompt,
+        provider,
+        test,
+        providerResponse: { output },
+      });
+      expect(result).toMatchObject({
+        pass: true,
+        reason: 'All assertions passed',
+      });
+    });
+
+    it('assert-set failure', async () => {
+      const output = 'Actual output';
+      const test: AtomicTestCase = {
+        assert: [
+          {
+            type: 'assert-set',
+            assert: [
+              {
+                type: 'equals',
+                value: 'Expected output',
+              },
+            ],
+          },
+        ],
+      };
+
+      const result: GradingResult = await runAssertions({
+        prompt,
+        provider,
+        test,
+        providerResponse: { output },
+      });
+      expect(result).toMatchObject({
+        pass: false,
+        reason: 'Expected output "Actual output" to equal "Expected output"',
+      });
+    });
+
+    it('assert-set threshold success', async () => {
+      const output = 'Expected output';
+      const test: AtomicTestCase = {
+        assert: [
+          {
+            type: 'assert-set',
+            threshold: 0.25,
+            assert: [
+              {
+                type: 'equals',
+                value: 'Hello world',
+                weight: 2,
+              },
+              {
+                type: 'contains',
+                value: 'Expected',
+                weight: 1,
+              },
+            ],
+          },
+        ],
+      };
+
+      const result: GradingResult = await runAssertions({
+        prompt,
+        provider,
+        test,
+        providerResponse: { output },
+      });
+      expect(result).toMatchObject({
+        pass: true,
+        reason: 'All assertions passed',
+      });
+    });
+
+    it('assert-set threshold failure', async () => {
+      const output = 'Expected output';
+      const test: AtomicTestCase = {
+        assert: [
+          {
+            type: 'assert-set',
+            threshold: 0.5,
+            assert: [
+              {
+                type: 'equals',
+                value: 'Hello world',
+                weight: 2,
+              },
+              {
+                type: 'contains',
+                value: 'Expected',
+                weight: 1,
+              },
+            ],
+          },
+        ],
+      };
+
+      const result: GradingResult = await runAssertions({
+        prompt,
+        provider,
+        test,
+        providerResponse: { output },
+      });
+      expect(result).toMatchObject({
+        pass: false,
+        reason: 'Aggregate score 0.33 < 0.5 threshold',
+      });
+    });
+
+    it('assert-set with metric', async () => {
+      const metric = 'The best metric';
+      const output = 'Expected output';
+      const test: AtomicTestCase = {
+        assert: [
+          {
+            type: 'assert-set',
+            metric,
+            threshold: 0.5,
+            assert: [
+              {
+                type: 'equals',
+                value: 'Hello world',
+              },
+              {
+                type: 'contains',
+                value: 'Expected',
+              },
+            ],
+          },
+        ],
+      };
+
+      const result: GradingResult = await runAssertions({
+        prompt,
+        provider,
+        test,
+        providerResponse: { output },
+      });
+      expect(result.namedScores).toStrictEqual({
+        [metric]: 0.5,
+      });
+    });
+
+    it('uses assert-set weight', async () => {
+      const output = 'Expected';
+      const test: AtomicTestCase = {
+        assert: [
+          {
+            type: 'equals',
+            value: 'Nope',
+            weight: 10,
+          },
+          {
+            type: 'assert-set',
+            weight: 90,
+            assert: [
+              {
+                type: 'equals',
+                value: 'Expected',
+              },
+            ],
+          },
+        ],
+      };
+
+      const result: GradingResult = await runAssertions({
+        prompt,
+        provider,
+        test,
+        providerResponse: { output },
+      });
+      expect(result.score).toBe(0.9);
+    });
+  });
+
+  it('preserves default provider', async () => {
+    const provider = new OpenAiChatCompletionProvider('gpt-4o-mini');
+    const output = 'Expected output';
+    const test: AtomicTestCase = {
+      assert: [
+        {
+          type: 'moderation',
+          provider: 'replicate:moderation:foo/bar',
+        },
+        {
+          type: 'llm-rubric',
+          value: 'insert rubric here',
+        },
+      ],
+    };
+
+    const callApiSpy = jest.spyOn(DefaultGradingJsonProvider, 'callApi').mockResolvedValue({
+      output: JSON.stringify({ pass: true, score: 1.0, reason: 'I love you' }),
+    });
+    const callModerationApiSpy = jest
+      .spyOn(ReplicateModerationProvider.prototype, 'callModerationApi')
+      .mockResolvedValue({ flags: [] });
+
+    const result: GradingResult = await runAssertions({
+      prompt: 'foobar',
+      provider,
+      test,
+      providerResponse: { output },
+    });
+
+    expect(result.pass).toBeTruthy();
+    expect(callApiSpy).toHaveBeenCalledTimes(1);
+    expect(callModerationApiSpy).toHaveBeenCalledTimes(1);
+  });
+});
